+++ products.php
@@ -73,7 +73,7 @@ $logUser = $_SESSION['user'];
             echo "<td>" . $row['status'] . "</td>";
             echo "<td><a href='view-product.php?id=" . $row['id'] . "'><i class=' typcn typcn-eye '></i></a> | ";
             echo "<a href='edit-product.php?id=" . $row['id'] . "&type=" . $row['typeid'] . "'><i class='typcn typcn-edit'></i></a> | ";
-            echo "<a href='delete-product.php?id=" . $row['id'] . "' class='text-danger'><i class='  typcn typcn-trash'></a></td>";
+            echo "<a href='delete-product.php?id=" . $row['id'] . "' class='text-danger onclick='return confirmDelete()'><i class='  typcn typcn-trash'></a></td>";
             echo "</tr>";
         }
 
@@ -91,6 +91,6 @@ include('footer.php');
 ?>
 <script>
 function confirmDelete() {
-    return confirm("Are you sure you want to delete this order?");
+    return confirm("Are you sure you want to delete this product?");
 }
 </script>
